Move agent structs to separate file
