+++ pkg/virt-launcher/virtwrap/agent/exec.go
@@ -9,31 +9,6 @@ import (
 	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/cli"
 )
 
-type execReturn struct {
-	Return execReturnData `json:"return"`
-}
-type execReturnData struct {
-	Pid int `json:"pid"`
-}
-
-type execStatusReturn struct {
-	Return execStatusReturnData `json:"return"`
-}
-type execStatusReturnData struct {
-	Exited   bool   `json:"exited"`
-	ExitCode int    `json:"exitcode"`
-	OutData  string `json:"out-data"`
-}
-
-// ExecExitCode returned at non-zero return codes
-type ExecExitCode struct {
-	ExitCode int
-}
-
-func (e ExecExitCode) Error() string {
-	return fmt.Sprint("exited with error code:", e.ExitCode)
-}
-
 // GuestExec sends the provided command and args to the guest agent for execution and returns an error on an unsucessful exit code
 // The resulting stdout will be returned as a string
 func GuestExec(virConn cli.Connection, domName string, command string, args []string, timeoutSeconds int32) (string, error) {
