+++ pkg/virt-launcher/virtwrap/agent/types.go
@@ -1,0 +1,28 @@
+package agent
+
+import "fmt"
+
+type execReturn struct {
+	Return execReturnData `json:"return"`
+}
+type execReturnData struct {
+	Pid int `json:"pid"`
+}
+
+type execStatusReturn struct {
+	Return execStatusReturnData `json:"return"`
+}
+type execStatusReturnData struct {
+	Exited   bool   `json:"exited"`
+	ExitCode int    `json:"exitcode"`
+	OutData  string `json:"out-data"`
+}
+
+// ExecExitCode returned at non-zero return codes
+type ExecExitCode struct {
+	ExitCode int
+}
+
+func (e ExecExitCode) Error() string {
+	return fmt.Sprint("exited with error code:", e.ExitCode)
+}
